Create Ant build file to make the customized jar process for lint rules easier
In order to run a custom lint rule, the java file, issue registry file,
and manifest file must be combined into a jar located in ~/.android/lint
directory. The jar must be using the custom manifest file, not the
default MANIFEST.MF. This is not easy to do using eclipse's export
wizard, so a custom Ant file is the solution.